backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.1.0
Flask-Migrate==4.0.0
Authlib==1.0.0
requests==2.28.2

# Optional, if using CORS for frontend-backend communication
# Flask-Cors==5.0.1

# Optional, for password hashing (if not using bcrypt or other libraries)
# Werkzeug==2.2.2

# For testing
# pytest==7.3.1
# pytest-flask==1.2.0

# For Docker (if you want to run the app in containers)
# gunicorn==20.1.0

# Database (if using Postgres)
# psycopg2==2.9.3

# Optional, for logging
# Flask-Logging==1.0.0
